+++ ForceSensor.cpp
@@ -10,7 +10,7 @@ void ForceSensor::read()
 {
     // https://learn.sparkfun.com/tutorials/force-sensitive-resistor-hookup-guide/all
 
-    double force = 0;
+    float force = 0;
 
     int raw_adc_reading = analogRead(pin);
 
@@ -19,15 +19,15 @@ void ForceSensor::read()
     if (raw_adc_reading != 0) // If the analog reading is non-zero
     {
         // Use ADC reading to calculate voltage:
-        double force_voltage = raw_adc_reading * M_VCC / 1023.0;
+        float force_voltage = raw_adc_reading * M_VCC / 1023.0;
 
         // Use voltage and static resistor value to
         // calculate FSR resistance:
-        double force_resistance = resistor_resistance * (M_VCC / force_voltage - 1.0);
+        float force_resistance = resistor_resistance * (M_VCC / force_voltage - 1.0);
 
         // Guesstimate force based on slopes in figure 3 of
         // FSR datasheet:
-        double force_conductance = 1.0 / force_resistance; // Calculate conductance
+        float force_conductance = 1.0 / force_resistance; // Calculate conductance
 
         // Break parabolic curve down into two linear slopes:
         if (force_resistance <= 600)
@@ -36,7 +36,7 @@ void ForceSensor::read()
             force = force_conductance / 0.000000642857;
     }
 
-    double processed = kalmanFilter.updateEstimate(force);
+    float processed = kalmanFilter.updateEstimate(force);
 
     *value = processed;
 }
